DWIM header insertion
Addtion of `gemini-insert-header-auto` which will insert a "do what I mean"
header that matches the current indentation level of headers preceeding it.

Supports creating a header from an active region, or the contents of the current
line if no region is active.

A smattering of "helper" functions have been added to aid in the insertion of
the header.